requirements.txt
eel==0.16.0
pymongo==4.6.1
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
bcrypt==4.1.2
python-jose==3.3.0
datetime
uuid 